+++ envelope/__init__.py
@@ -1,6 +1,9 @@
 import sys
-from .envelope import Envelope
 
-sys.modules[__name__] = Envelope  # to allow dynamic module call `import envelope; envelope(...)`
-envelope = Envelope  # to support static inspection autocompletion `from envelope import envelope`
-Envelope.envelope = Envelope  # point `from envelope import envelope` at working `import envelope` at runtime
+from .envelope import envelope
+
+# __all__ = [Envelope]
+
+sys.modules[__name__] = envelope  # to allow dynamic module call `import envelope; envelope(...)`
+envelope = envelope  # to support static inspection autocompletion `from envelope import envelope`
+envelope.envelope = envelope  # point `from envelope import envelope` at working `import envelope` at runtime